Move common functions to an include file
Move three common functions to a single file rather than redefining
them in each file.

We use a shutdown function in output_header() to automatically call
output_footer() when the PHP script ends. This ensures that the HTML
page is always properly closed and avoids the need to call the footer
function at the end of every page.