Use subdirs under bin/ for external tools
For pptext and pphtml which come from external repositories, place them
under subdirectories in bin/ such that they can simply be cloned. This
will make it much easier to automate updates separately for ppwb,
pptext, and pphtml.

Use this manual checkout process instead of git submodules because the
latter are going to be too heavy-handed (and possibly a bit obscure to
maintainers) for what is needed here.