Add support for ppcomp 'suppress-sidenote-tags'
PPWB already supports the css-add-sidenote feature of ppcomp, which adds
a "[Sidenote: ]" wrapper to the HTML file where the .sidenote class is
applied.

This commit adds support for the other case: removing the "[Sidenote: ]"
from the text file. This can be useful because the .sidenote class may
not be the way sidenotes were implemented (i.e. DPwiki outlines
suggested methods for inline sidenote markup using .sni, in which case
the .sidenote class would not appear and ppcomp would not "do the right
thing" for this case.